remarks correction
