+++ services/adgs/rs_server_adgs/adgs_utils.py
@@ -25,7 +25,7 @@ from pathlib import Path
 import stac_pydantic
 import yaml
 from fastapi import HTTPException, status
-from rs_server_common.stac_api_common import map_stac_platform
+from rs_server_common.utils.utils import map_stac_platform
 
 ADGS_CONFIG = Path(osp.realpath(osp.dirname(__file__))).parent / "config"
 search_yaml = ADGS_CONFIG / "adgs_search_config.yaml"
@@ -84,47 +84,6 @@ def serialize_adgs_asset(feature_collection, products):
         feature.assets[feature.id] = feature.assets.pop("file")
         feature.id = feature.id.rsplit(".", 1)[0]  # remove extension if any
     return feature_collection
-
-
-def auxip_map_mission(platform: str, constellation: str) -> tuple[str | None, str | None]:
-    """
-    Custom function for ADGS, to read constellation mapper and return propper
-    values for platform and serial.
-    Eodag maps this values to platformShortName, platformSerialIdentifier
-
-    Input: platform = sentinel-1a       Output: sentinel-1, A
-    Input: platform = sentinel-5P       Output: sentinel-5p, None
-    Input: constellation = sentinel-1   Output: sentinel-1, None
-    """
-    data = map_stac_platform()
-    platform_short_name: str | None = None
-    platform_serial_identifier: str | None = None
-    try:
-        if platform:
-            config = next(satellite[platform] for satellite in data["satellites"] if platform in satellite)
-            platform_short_name = config.get("constellation", None)
-            platform_serial_identifier = config.get("serialid", None)
-        if constellation:
-            if platform_short_name and platform_short_name != constellation:
-                # Inconsistent combination of platform / constellation case
-                raise HTTPException(
-                    status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-                    detail="Invalid combination of platform-constellation",
-                )
-            if any(
-                satellite[list(satellite.keys())[0]]["constellation"] == constellation
-                for satellite in data["satellites"]
-            ):
-                platform_short_name = constellation
-                platform_serial_identifier = None
-            else:
-                raise KeyError
-    except (KeyError, IndexError, StopIteration) as exc:
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="Cannot map platform/constellation",
-        ) from exc
-    return platform_short_name, platform_serial_identifier
 
 
 def adgs_reverse_map_mission(
