+++ services/prip/rs_server_prip/prip_utils.py
@@ -30,7 +30,7 @@ from typing import Any
 import stac_pydantic
 import yaml
 from fastapi import HTTPException, status
-from rs_server_common.stac_api_common import map_stac_platform
+from rs_server_common.utils.utils import map_stac_platform
 
 PRIP_CONFIG = Path(osp.realpath(osp.dirname(__file__))).parent / "config"
 search_yaml = PRIP_CONFIG / "prip_search_config.yaml"
@@ -118,50 +118,6 @@ def serialize_prip_asset(feature_collection: stac_pydantic.ItemCollection, produ
         feature.id = new_key
 
     return feature_collection
-
-
-# ----------------------
-# Platform mapping utilities
-# ----------------------
-def prip_map_mission(platform: str, constellation: str) -> tuple[str | None, str | None]:
-    """
-    Custom function for PRIP, to read constellation mapper and return propper
-    values for platform and serial.
-    Eodag maps this values to platformShortName, platformSerialIdentifier
-
-    Input: platform = sentinel-1a       Output: sentinel-1, A
-    Input: platform = sentinel-5P       Output: sentinel-5p, None
-    Input: constellation = sentinel-1   Output: sentinel-1, None
-    """
-    data = map_stac_platform()
-    platform_short_name: str | None = None
-    platform_serial_identifier: str | None = None
-    try:
-        if platform:
-            config = next(satellite[platform] for satellite in data["satellites"] if platform in satellite)
-            platform_short_name = config.get("constellation", None)
-            platform_serial_identifier = config.get("serialid", None)
-        if constellation:
-            if platform_short_name and platform_short_name != constellation:
-                # Inconsistent combination of platform / constellation case
-                raise HTTPException(
-                    status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-                    detail="Invalid combination of platform-constellation",
-                )
-            if any(
-                satellite[list(satellite.keys())[0]]["constellation"] == constellation
-                for satellite in data["satellites"]
-            ):
-                platform_short_name = constellation
-                platform_serial_identifier = None
-            else:
-                raise KeyError
-    except (KeyError, IndexError, StopIteration) as exc:
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="Cannot map platform/constellation",
-        ) from exc
-    return platform_short_name, platform_serial_identifier
 
 
 def prip_reverse_map_mission(
